huge changes to database and refactored tests to compliment this change

+++ migrations/005.do.create-room_queue.sql
@@ -2,7 +2,7 @@ begin;
 
 create table if not exists room_queue (
     id integer default 1,
-    size integer default 0,
+    size smallint default 0,
     first integer references game_history(id) default null,
     last integer references game_history(id) default null
 );
